Moves the 'consent' command to the 'cli' namespace. Closes #1336
