+++ src/o365/cli/commands.ts
@@ -5,5 +5,6 @@ export default {
   COMPLETION_PWSH_SETUP: `${prefix} completion pwsh setup`,
   COMPLETION_PWSH_UPDATE: `${prefix} completion pwsh update`,
   COMPLETION_SH_SETUP: `${prefix} completion sh setup`,
-  COMPLETION_SH_UPDATE: `${prefix} completion sh update`
+  COMPLETION_SH_UPDATE: `${prefix} completion sh update`,
+  CONSENT: `${prefix} consent`
 };
